Harden prepared SVD contracts after review
Why: alias checks, feature boundaries, and validation order must remain safe and independently testable before callers rely on reusable factorization state.
